# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OsCFinder API",
    "description": "Internal API for OsCFinder, a B2B lead-generation SaaS. Generated directly from the route handlers under app/api/** — every schema here reflects what the code actually reads/returns, not a hand-written guess.",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/", "description": "Same origin as the web app" }],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Session-based auth via Supabase Auth cookies set after logging in through the web app (see lib/auth.ts: getSession()/requireAuth()/requireAdmin()/requireActiveAccount()). There is no API key or bearer token to pass manually for these routes — they are only callable from an authenticated browser session."
      },
      "cronSecret": {
        "type": "http",
        "scheme": "bearer",
        "description": "Used ONLY by the /api/campaigns/process worker route. Requires header `Authorization: Bearer <CRON_SECRET>` matching the CRON_SECRET env var (see app/api/campaigns/process/route.ts). Entirely separate from user session auth above — a valid user cookie does not authorize this route, and this bearer token does not authorize any other route."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"]
      },
...
      "Company": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "industry": { "type": "string", "nullable": true },
          "location": { "type": "string", "nullable": true },
          "plan": { "type": "string", "enum": ["starter", "business", "enterprise", "demo"] },
          "status": { "type": "string", "enum": ["inactive", "active", "suspended", "churned"] },
          "setup_fee_paid": { "type": "boolean" },
          "renewal_fee_paid": { "type": "boolean" },
          "plan_start_date": { "type": "string", "format": "date-time" },
          "plan_end_date": { "type": "string", "format": "date-time" },
          "is_demo": { "type": "boolean" },
          "demo_expires_at": { "type": "string", "format": "date-time", "nullable": true },
          "demo_converted": { "type": "boolean" },
          "demo_notes": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/admin/companies": {
      "get": {
        "tags": ["Admin"],
        "summary": "List every company with this-month usage, from the admin_company_overview view.",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object", "additionalProperties": true } } } } },
          "403": { "description": "Forbidden -- not an admin", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Create a company, its Supabase Auth user, and its users table record.",
        "security": [{ "cookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "plan": { "type": "string", "enum": ["starter", "business", "enterprise"], "default": "starter" },
                  "password": { "type": "string" },
                  "full_name": { "type": "string" },
                  "industry": { "type": "string" },
                  "location": { "type": "string" },
                  "setup_fee_paid": { "type": "boolean", "default": false },
                  "plan_start_date": { "type": "string", "format": "date-time" },
                  "plan_end_date": { "type": "string", "format": "date-time" },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Created", "content": { "application/json": { "schema": { "type": "object", "properties": { "company": { "$ref": "#/components/schemas/Company" }, "user_id": { "type": "string", "format": "uuid" } } } } } },
          "400": { "description": "Missing required fields, or invalid plan", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "description": "Forbidden -- not an admin", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}